chore: Rename `jobs` package to `jobq`
