+++ client/src/jobq/__init__.py
@@ -1,6 +1,6 @@
-from jobs import assembler
-from jobs.image import Image
-from jobs.job import (
+from jobq import assembler
+from jobq.image import Image
+from jobq.job import (
     ImageOptions,
     Job,
     JobOptions,
