+++ client/src/jobq/utils/math.py
